\name{calcSumHoldings}
\alias{calcSumHoldings}
\title{
Calculates Summary of Portfolio Holdings
}
\description{
Aggregates all of stocks value in portfolio holdings group by portfolio ID
}
\usage{
calcSumHoldings(histMarketData, holdingsData)
}
\arguments{
  \item{histMarketData}{
Historical Market Data
}
  \item{holdingsData}{
Portfolio Holding Data
}
}

\examples{
## see demo.R
}
