\name{calendarHeatMap}
\alias{calendarHeatMap}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Plots Calendar Heat Map
}
\description{
Plots Calendar Heat Map of a stock from our portfolio DB or from web services
}
\usage{
calendarHeatMap(mData, ticker, startdate = "2009-08-21", enddate = "2010-08-20")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mData}{
Market Data
}
  \item{ticker}{
Ticker symbol
}
  \item{startdate}{
start date
}
  \item{enddate}{
end date
}
}
\references{
Wickham, Hadley (2009) ggplot2
Chambers, J. (2008) Software for Data Analysis
}
\author{
Tirto Adji
}
\examples{
## see demo.R
}
