\name{portfolioBoxPlot}
\alias{portfolioBoxPlot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Plots Stock Prices Box Plot
}
\description{
Plots Stock Prices Box Plot in our porfolio from low to high prices to identify outliers
}
\usage{
portfolioBoxPlot(mData, tData)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mData}{
Market Data
}
  \item{tData}{
Transaction Data
}
}
\references{
Wickham, Hadley (2009) ggplot2
Chambers, J. (2008) Software for Data Analysis
}

\examples{
## see demo.R
}
