\name{portfolioHoldingSummaryChart}
\alias{portfolioHoldingSummaryChart}
\title{
Plots Portfolio Holding Summary Chart
}
\description{
Compares performance of algorithms used in determining trade strategy in different portfolio holdings
}
\usage{
portfolioHoldingSummaryChart(hData, pData, xrng = NULL, yrng = NULL, caption = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{hData}{
Historical Stock Market Data
}
  \item{pData}{
Portfolio Data
}
  \item{xrng,yrng,caption}{
optional argument of x, y range and caption
}
}

\references{
Wickham, Hadley (2009) ggplot2
Chambers, J. (2008) Software for Data Analysis
}
\author{
Tirto Adji
}
\examples{
## see demo.R
}
