\name{portfolioStocksChart}
\alias{portfolioStocksChart}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Plots Portfolio Stocks Chart
}
\description{
Compares performance of different stock prices in our portfolio
}
\usage{
portfolioStocksChart(data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{
Portfolio Data
}
}
\references{
Wickham, Hadley (2009) ggplot2
Chambers, J. (2008) Software for Data Analysis
}
\author{
Tirto Adji
}
\examples{
# see demo.R
}
