\name{stockCorrMatrixChart}
\alias{stockCorrMatrixChart}
\title{
Plots stock correlation matrix chart
}
\description{
This function initialize and plots stock prices that are highly correlated
}
\usage{
stockCorrMatrixChart(data)
}
\arguments{
  \item{data}{
Correlation matrix data
}
}
\author{
Tirto Adji
}
\examples{
## see demo.R
}
