\name{stockCorrPairsCartesianChart}
\alias{stockCorrPairsCartesianChart}
\title{
Plots stock correlation cartesian chart
}
\description{
This function initializes and plots stock prices that are highly correlated
}
\usage{
stockCorrPairsCartesianChart(sMarketData,corrData)
}
\arguments{
  \item{sMarketData}{
Stock Market Data
}
  \item{corrData}{
Correlation matrix data
}
}
\author{
Tirto Adji
}
\examples{
## see demo.R
}
