\name{TradeStrategyVisualizer-package}
\alias{TradeStrategyVisualizer-package}
\alias{TradeStrategyVisualizer}
\docType{package}
\title{
Trade Strategy Visualizer
}
\description{
This package contains a trade simulator and a collection of trading strategies
}
\details{
\tabular{ll}{
Package: \tab TradeStrategyVisualizer\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-03-12\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Tirto Adji
}
\references{
Wickham, Hadley (2009) ggplot2
Chambers, J. (2008) Software for Data Analysis
}

\keyword{ package }

